Refactored comms / migrant collision methods
